+++ Core/Defaults.lua
@@ -89,6 +89,9 @@ addon.defaults = {
       sortItems = true, -- Sort sessions by item type and item level
       rejectTrade = false, -- Can candidates choose not to give loot to the council
       autoTrade = false,
+      awardLater = false, -- Auto check award later
+
+      chatFrameName = "DEFAULT_CHAT_FRAME", -- The chat frame to use for :Print()s
 
       UI = { -- stores all ui information
          ['**'] = { -- Defaults
@@ -175,6 +178,7 @@ addon.defaults = {
 
       enableHistory = true,
       sendHistory = true,
+      sendHistoryToGuildChannel = false,
 
       minRank = -1,
       council = {},